pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentbeats-swe-verified"
version = "0.1.0"
description = "Agentbeats SWE Verified"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "a2a>=0.44",
    "a2a-sdk>=0.3.5",
    "google-adk>=1.14.1",
    "google-genai>=1.36.0",
    "litellm>=1.79.3",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
    "pyarrow>=18.1.0",
    "pandas>=2.2.3",
    "loguru>=0.7.3",
]

[project.scripts]
agentbeats-run = "agentbeats.run_scenario:main"

[dev]
package = true
dependencies = [
    "mypy>=1.18.1",
]

[tool.hatch.build.targets.wheel]
packages = ["src/agentbeats"]
